task: add generic type to couple payload with the task
