+++ featurebyte/worker/registry.py
@@ -6,7 +6,7 @@ from typing import Dict, Type
 from enum import Enum
 
 from featurebyte.enum import WorkerCommand
-from featurebyte.worker.task.base import BaseTask
+from featurebyte.worker.task.base import BaseTask, TaskT
 from featurebyte.worker.task.batch_feature_create import BatchFeatureCreateTask
 from featurebyte.worker.task.batch_feature_table import BatchFeatureTableTask
 from featurebyte.worker.task.batch_request_table import BatchRequestTableTask
@@ -32,7 +32,7 @@ from featurebyte.worker.task.test_task import TestTask
 from featurebyte.worker.task.tile_task import TileTask
 
 # TASK_REGISTRY_MAP contains a mapping of the worker command to the task.
-TASK_REGISTRY_MAP: Dict[Enum, Type[BaseTask]] = {
+TASK_REGISTRY_MAP: Dict[Enum, Type[BaseTask[TaskT]]] = {  # type: ignore[valid-type]
     WorkerCommand.FEATURE_JOB_SETTING_ANALYSIS_CREATE: FeatureJobSettingAnalysisTask,
     WorkerCommand.FEATURE_JOB_SETTING_ANALYSIS_BACKTEST: FeatureJobSettingAnalysisBacktestTask,
     WorkerCommand.HISTORICAL_FEATURE_TABLE_CREATE: HistoricalFeatureTableTask,
