+++ featurebyte/worker/task/batch_request_table.py
@@ -3,14 +3,12 @@ BatchRequestTable creation task
 """
 from __future__ import annotations
 
-from typing import Any, cast
+from typing import Any
 
 from uuid import UUID
 
 from featurebyte.logging import get_logger
-from featurebyte.models.base import User
 from featurebyte.models.batch_request_table import BatchRequestTableModel
-from featurebyte.persistent import Persistent
 from featurebyte.schema.worker.task.batch_request_table import BatchRequestTableTaskPayload
 from featurebyte.service.batch_request_table import BatchRequestTableService
 from featurebyte.service.feature_store import FeatureStoreService
@@ -21,7 +19,7 @@ from featurebyte.worker.task.mixin import DataWarehouseMixin
 logger = get_logger(__name__)
 
 
-class BatchRequestTableTask(DataWarehouseMixin, BaseTask):
+class BatchRequestTableTask(DataWarehouseMixin, BaseTask[BatchRequestTableTaskPayload]):
     """
     BatchRequestTable Task
     """
@@ -31,34 +29,23 @@ class BatchRequestTableTask(DataWarehouseMixin, BaseTask):
     def __init__(  # pylint: disable=too-many-arguments
         self,
         task_id: UUID,
-        payload: dict[str, Any],
         progress: Any,
-        user: User,
-        persistent: Persistent,
         feature_store_service: FeatureStoreService,
         session_manager_service: SessionManagerService,
         batch_request_table_service: BatchRequestTableService,
     ):
         super().__init__(
             task_id=task_id,
-            payload=payload,
             progress=progress,
-            user=user,
-            persistent=persistent,
         )
         self.feature_store_service = feature_store_service
         self.session_manager_service = session_manager_service
         self.batch_request_table_service = batch_request_table_service
 
-    async def get_task_description(self) -> str:
-        payload = cast(BatchRequestTableTaskPayload, self.payload)
+    async def get_task_description(self, payload: BatchRequestTableTaskPayload) -> str:
         return f'Save batch request table "{payload.name}"'
 
-    async def execute(self) -> Any:
-        """
-        Execute BatchRequestTable task
-        """
-        payload = cast(BatchRequestTableTaskPayload, self.payload)
+    async def execute(self, payload: BatchRequestTableTaskPayload) -> Any:
         feature_store = await self.feature_store_service.get_document(
             document_id=payload.feature_store_id
         )
@@ -73,13 +60,13 @@ class BatchRequestTableTask(DataWarehouseMixin, BaseTask):
             sample_rows=None,
         )
 
-        async with self.drop_table_on_error(db_session, location.table_details, self.payload):
+        async with self.drop_table_on_error(db_session, location.table_details, payload):
             columns_info, num_rows = await service.get_columns_info_and_num_rows(
                 db_session, location.table_details
             )
             logger.debug("Creating a new BatchRequestTable", extra=location.table_details.dict())
             batch_request_table = BatchRequestTableModel(
-                _id=self.payload.output_document_id,
+                _id=payload.output_document_id,
                 user_id=payload.user_id,
                 name=payload.name,
                 location=location,
