+++ featurebyte/worker/task/historical_feature_table.py
@@ -3,14 +3,12 @@ HistoricalFeatureTable creation task
 """
 from __future__ import annotations
 
-from typing import Any, cast
+from typing import Any
 
 from uuid import UUID
 
 from featurebyte.logging import get_logger
-from featurebyte.models.base import User
 from featurebyte.models.historical_feature_table import HistoricalFeatureTableModel
-from featurebyte.persistent import Persistent
 from featurebyte.schema.worker.task.historical_feature_table import (
     HistoricalFeatureTableTaskPayload,
 )
@@ -21,11 +19,12 @@ from featurebyte.service.session_manager import SessionManagerService
 from featurebyte.worker.task.base import BaseTask
 from featurebyte.worker.task.mixin import DataWarehouseMixin
 from featurebyte.worker.util.observation_set_helper import ObservationSetHelper
+from featurebyte.worker.util.task_progress_updater import TaskProgressUpdater
 
 logger = get_logger(__name__)
 
 
-class HistoricalFeatureTableTask(DataWarehouseMixin, BaseTask):
+class HistoricalFeatureTableTask(DataWarehouseMixin, BaseTask[HistoricalFeatureTableTaskPayload]):
     """
     HistoricalFeatureTableTask creates a HistoricalFeatureTable by computing historical features
     """
@@ -35,38 +34,29 @@ class HistoricalFeatureTableTask(DataWarehouseMixin, BaseTask):
     def __init__(  # pylint: disable=too-many-arguments
         self,
         task_id: UUID,
-        payload: dict[str, Any],
         progress: Any,
-        user: User,
-        persistent: Persistent,
         feature_store_service: FeatureStoreService,
         session_manager_service: SessionManagerService,
         observation_set_helper: ObservationSetHelper,
         historical_feature_table_service: HistoricalFeatureTableService,
         historical_features_service: HistoricalFeaturesService,
+        task_progress_updater: TaskProgressUpdater,
     ):
         super().__init__(
             task_id=task_id,
-            payload=payload,
             progress=progress,
-            user=user,
-            persistent=persistent,
         )
         self.feature_store_service = feature_store_service
         self.session_manager_service = session_manager_service
         self.observation_set_helper = observation_set_helper
         self.historical_feature_table_service = historical_feature_table_service
         self.historical_features_service = historical_features_service
+        self.task_progress_updater = task_progress_updater
 
-    async def get_task_description(self) -> str:
-        payload = cast(HistoricalFeatureTableTaskPayload, self.payload)
+    async def get_task_description(self, payload: HistoricalFeatureTableTaskPayload) -> str:
         return f'Save historical feature table "{payload.name}"'
 
-    async def execute(self) -> Any:
-        """
-        Execute HistoricalFeatureTableTask
-        """
-        payload = cast(HistoricalFeatureTableTaskPayload, self.payload)
+    async def execute(self, payload: HistoricalFeatureTableTaskPayload) -> Any:
         feature_store = await self.feature_store_service.get_document(
             document_id=payload.feature_store_id
         )
@@ -80,13 +70,13 @@ class HistoricalFeatureTableTask(DataWarehouseMixin, BaseTask):
         )
 
         async with self.drop_table_on_error(
-            db_session=db_session, table_details=location.table_details, payload=self.payload
+            db_session=db_session, table_details=location.table_details, payload=payload
         ):
             await self.historical_features_service.compute(
                 observation_set=observation_set,
                 compute_request=payload.featurelist_get_historical_features,
                 output_table_details=location.table_details,
-                progress_callback=self.update_progress,
+                progress_callback=self.task_progress_updater.update_progress,
             )
             (
                 columns_info,
@@ -99,7 +89,7 @@ class HistoricalFeatureTableTask(DataWarehouseMixin, BaseTask):
             )
             historical_feature_table = HistoricalFeatureTableModel(
                 _id=payload.output_document_id,
-                user_id=self.payload.user_id,
+                user_id=payload.user_id,
                 name=payload.name,
                 location=location,
                 observation_table_id=payload.observation_table_id,
