+++ featurebyte/worker/task/observation_table_upload.py
@@ -3,17 +3,14 @@ ObservationTable creation task
 """
 from __future__ import annotations
 
-from typing import Any, cast
+from typing import Any
 
 from pathlib import Path
 from uuid import UUID
 
 from featurebyte.logging import get_logger
-from featurebyte.models.base import User
 from featurebyte.models.observation_table import ObservationTableModel, UploadedFileInput
 from featurebyte.models.request_input import RequestInputType
-from featurebyte.persistent import Persistent
-from featurebyte.schema.worker.task.observation_table import ObservationTableTaskPayload
 from featurebyte.schema.worker.task.observation_table_upload import (
     ObservationTableUploadTaskPayload,
 )
@@ -27,7 +24,7 @@ from featurebyte.worker.task.mixin import DataWarehouseMixin
 logger = get_logger(__name__)
 
 
-class ObservationTableUploadTask(DataWarehouseMixin, BaseTask):
+class ObservationTableUploadTask(DataWarehouseMixin, BaseTask[ObservationTableUploadTaskPayload]):
     """
     ObservationTableUpload Task
     """
@@ -37,10 +34,7 @@ class ObservationTableUploadTask(DataWarehouseMixin, BaseTask):
     def __init__(  # pylint: disable=too-many-arguments
         self,
         task_id: UUID,
-        payload: dict[str, Any],
         progress: Any,
-        user: User,
-        persistent: Persistent,
         temp_storage: Storage,
         feature_store_service: FeatureStoreService,
         session_manager_service: SessionManagerService,
@@ -48,25 +42,17 @@ class ObservationTableUploadTask(DataWarehouseMixin, BaseTask):
     ):
         super().__init__(
             task_id=task_id,
-            payload=payload,
             progress=progress,
-            user=user,
-            persistent=persistent,
         )
         self.temp_storage = temp_storage
         self.feature_store_service = feature_store_service
         self.session_manager_service = session_manager_service
         self.observation_table_service = observation_table_service
 
-    async def get_task_description(self) -> str:
-        payload = cast(ObservationTableTaskPayload, self.payload)
+    async def get_task_description(self, payload: ObservationTableUploadTaskPayload) -> str:
         return f'Upload observation table "{payload.name}" from CSV.'
 
-    async def execute(self) -> Any:
-        """
-        Execute ObservationTableUpload task
-        """
-        payload = cast(ObservationTableUploadTaskPayload, self.payload)
+    async def execute(self, payload: ObservationTableUploadTaskPayload) -> Any:
         feature_store = await self.feature_store_service.get_document(
             document_id=payload.feature_store_id
         )
@@ -85,7 +71,7 @@ class ObservationTableUploadTask(DataWarehouseMixin, BaseTask):
         # Write the file to the warehouse
         await db_session.register_table(location.table_details.table_name, uploaded_dataframe)
 
-        async with self.drop_table_on_error(db_session, location.table_details, self.payload):
+        async with self.drop_table_on_error(db_session, location.table_details, payload):
             # Validate table and retrieve metadata about the table
             additional_metadata = (
                 await self.observation_table_service.validate_materialized_table_and_get_metadata(
@@ -98,7 +84,7 @@ class ObservationTableUploadTask(DataWarehouseMixin, BaseTask):
             # Store metadata of the observation table in mongo
             logger.debug("Creating a new ObservationTable", extra=location.table_details.dict())
             observation_table = ObservationTableModel(
-                _id=self.payload.output_document_id,
+                _id=payload.output_document_id,
                 user_id=payload.user_id,
                 name=payload.name,
                 location=location,
