+++ tests/unit/worker/task/test_batch_feature_table.py
@@ -21,6 +21,5 @@ async def test_get_task_description(catalog, app_container):
         deployment_id=ObjectId(),
         catalog_id=catalog.id,
     )
-    app_container.override_instance_for_test("payload", payload.dict(by_alias=True))
     task = app_container.get(BatchFeatureTableTask)
-    assert await task.get_task_description() == 'Save batch feature table "Test Features"'
+    assert await task.get_task_description(payload) == 'Save batch feature table "Test Features"'
