+++ tests/unit/worker/task/test_feature_job_setting_analysis.py
@@ -229,12 +229,11 @@ class TestFeatureJobSettingAnalysisTask(BaseTaskTestSuite):
         """
         Test get task description
         """
-        payload = FeatureJobSettingAnalysisTask.payload_class(**self.payload)
         app_container.override_instance_for_test("catalog_id", catalog.id)
         app_container.override_instance_for_test("persistent", persistent)
-        app_container.override_instance_for_test("payload", payload.dict(by_alias=True))
         task = app_container.get(FeatureJobSettingAnalysisTask)
+        payload = task.get_payload_obj(self.payload)
         assert (
-            await task.get_task_description()
+            await task.get_task_description(payload)
             == 'Analyze feature job settings for table "sf_event_table"'
         )
