+++ tests/unit/worker/task/test_online_store_cleanup.py
@@ -19,8 +19,8 @@ async def test_get_task_description(app_container):
         feature_store_id=ObjectId(),
         online_store_table_name="Test Online Store Table",
     )
-    app_container.override_instance_for_test("payload", payload.dict(by_alias=True))
     task = app_container.get(OnlineStoreCleanupTask)
     assert (
-        await task.get_task_description() == 'Clean up online store table "Test Online Store Table"'
+        await task.get_task_description(payload)
+        == 'Clean up online store table "Test Online Store Table"'
     )
