+++ tests/unit/worker/task/test_target_table.py
@@ -32,6 +32,5 @@ async def test_get_task_description(catalog, app_container):
             ),
         ),
     )
-    app_container.override_instance_for_test("payload", payload.dict(by_alias=True))
     task = app_container.get(TargetTableTask)
-    assert await task.get_task_description() == 'Save target table "Test Target Table"'
+    assert await task.get_task_description(payload) == 'Save target table "Test Target Table"'
